Fix share button display
With ownCloud 8 the "Close" button of the PDF Viewer is shown incorrectly and overlaps the "secondary tool bar" button  which results in a bad user experience and looks very ugly.

This patch addresses this problem by moving the "Close" button inside the iframe and accessing it using jQuery.

Testplan:

- [ ] Accessing a PDF as logged-in user from the file list works and a Close button is shown.
  - [ ] Clicking on the close button closes the file
- [ ] Sharing a single PDF file via link works and no Close button is shown
- [ ] Sharing a folder containing a PDF file via link works and a Close button is shown.
   - [ ] Clicking on the close button closes the file

To test this properly you need probably to checkout https://github.com/owncloud/core/pull/14773