Set custom initialization for PDF.js
In PDF.js 1.9.426 it is no longer possible to simply set the custom
values for "PDFJS" and "PDFViewerApplication" when "js/workersrc.js" is
parsed. When "PDFViewerApplication.initialize" is executed it overwrites
the value of "PDFJS.workerSrc", so the custom initialization has to be
executed after that. As "PDFViewerApplication" does not provide any
hookable point for custom initialization in its initialization routine a
dirty hack has to be used.

When "vendor/pdfjs/web/viewer.js" is parsed at the end it calls
"PDFViewerApplication.initialize" (either directly or as a
"DOMContentLoaded" event handler), and "PDFViewerApplication.initialize"
sends an asynchronous XHR request to initialize the locales, which
causes the execution flow to continue with the next script or the next
"DOMContentLoaded" handler. Thanks to this
"initializeCustomPDFViewerApplication" can be executed at that point by
parsing "js/workersrc.js" after "vendor/pdfjs/web/viewer.js" and either
calling it directly or adding it as a "DOMContentLoaded" event handler
(using the same conditions as in "vendor/pdfjs/web/viewer.js").

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>