Disable "Open file" button
Due to the changes from PDF.js 1.4.20 to 1.9.426 now trying to open a
different PDF file fails with a Content Security Policy violation.
Instead of fixing that the "Open file" button was disabled, as being
able to open a different PDF file from the local file system to view it
instead of the PDF file in the server did not provide much value.

The button was disabled by setting the "hidden" attribute to "true"
instead of with the CSS class for consistency with how it is done in
"vendor/pdfjs/web/viewer.js" when the browser does not support any of
the APIs needed to open the local file.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>