Enable full screen mode for PDF files
The PDF viewer provides a "Full screen" button when it detects that the
browser supports the full screen mode and that the PDF is inside an
element that can be displayed full screen. Therefore, as the PDF viewer
is shown inside an "iframe" element its "allowfullscreen" attribute must
be set to "true".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>