Use download manager from PDF.js to download the file
The default implementation of "PDFViewerApplication.download" creates a
blob which is then downloaded using the download manager from PDF.js.
This causes a CSP error in Firefox, so in 7bcdc4fb29 the default
implementation was replaced with a "window.open" call with the URL of
the file.

Besides through a blob, the download manager from PDF.js also supports
downloading a file through its URL. Thus, in order to keep closer to the
default PDF.js viewer, now the custom implementation of
"PDFViewerApplication.download" was changed to download the file using
the download manager from PDF.js.

Note that "window.open" does not work in Internet Explorer 11 (in this
specific case, probably due to the use of iframes or something like
that), so this change also fixes the download in Internet Explorer 11,
it is not just an "academic" change.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>