Inherit content padding to show the PDF frame below the header
The PDF frame is a direct child of "#content" that fills the whole
element. This is done using absolute positioning, so the containing box
for the PDF frame is the padding box of "#content". Thus, the same
padding used in "#content" to show it below the header must be used in
the PDF frame too (otherwise the PDF frame would appear behind the
header).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>