Fix embedded viewer with new server layout on IE 11
The embedded PDF viewer is shown in an iframe in front of the of the
Files app by using absolute positioning. Since Nextcloud 14 the parent
element of the iframe uses a flex layout, and the previous sibling of
the iframe, the "app-content" element, is shown with
"flex-basis: 100vw".

In IE 11 the scenario above causes the iframe to be shown to the right
of "app-content" instead of in front of it; it is necessary to
explicitly set "left: 0" to force the iframe to be aligned to the left
of its positioning context (which in this case is its parent element)
and thus in front of "app-content".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>