Hide footer in public share page
The footer in the public share page is hidden by adding the "hidden" CSS
class to the footer. However, the CSS rules for the footer defined in
the server are more specific than the general rule for ".hidden"
elements, which caused the footer to be shown even if marked as
".hidden". Now a more specific rule was added to override the ones from
the server and ensure that the footer is hidden.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>