Fix load of character maps
The character maps are used by PDF.JS for font rendering, and are needed
to properly show some texts (for example, Japanese text). The PDF.JS
package provides the needed character maps, but the URL to get them must
be set with the "PDFJS.cMapUrl" property in the same way as the
"PDFJs.workerSrc" property is set (as the default value set in
"viewer.js" is not right for the PDF viewer app).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>