Replace PDF.js v2.0.943 files with v2.1.266 files
The example PDF ("compressed.tracemonkey-pldi-09.pdf") is not needed and
thus was not included; "templates/viewer.php" acts as "viewer.html", so
the later was not included either.

On the other hand, source maps were kept so they could be used in case
of need.

Note that "viewer.html" did not change between v2.0.943 and v2.1.266, so
there is no need to update "templates/viewer.php".

Also note that, unlike v2.0.943, PDF.js v2.1.266 works out of the box
with Nextcloud Content Security Policy.

Signed-off-by: Julius Härtl <jus@bitgrid.net>