Update custom PDFViewerApplication initialization to PDF.js 2.0.943
The PDFViewerApplication options are no longer got from the PDFJS object
but from a helper object that holds the configuration. This object is
globally set before "PDFViewerApplication.initialize" is called, so it
can be used in the custom initialization code.

Note that despite setting "isEvalSupported" to false this does not work
out of the box with the default Nextcloud Content Security Policy. Due
to how the PDF.js library is bundled it tries to execute an "eval"
function, which is blocked by the CSP and causes the viewer to not load
the PDF file.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>