Update custom PDFViewerApplication initialization to PDF.js 2.1.266
Since v2.1.266 an event is triggered before the PDFViewerApplication is
initialized, so the hack that relied on the asynchronous calls during
the initialization to hook in is no longer needed.

However, as the event is triggered before the initialization the options
are now set before the preferences are read. As the preferences override
the options the preferences need to be disabled to be able to set
"externalLinkTarget". Disabling the preferences has no effect for the
rest of them, as they already have the same value as the default
options.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>