Prevent text selection when download is hidden
When the PDF is rendered PDFViewer creates canvases with the contents of
the pages and, in front of them, it creates too a layer with HTML text
elements that makes possible to select it. That text layer is used too
by the search function to highlight the results, so it can not be
removed. Instead of that the selection of the text is prevented using
CSS.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>