Prevent printing through shortcuts when download is hidden
Even if the Print button was hidden in the toolbar it was still possible
to trigger the printing by using Ctrl|Meta+P. This uses the print
service from PDFViewer, so when the download is hidden the print service
is now disabled.

When printing is flagged as not supported in the PDFViewerApplication
initialization the Print buttons are disabled, so it is no longer needed
to disable them explicity. Moreover, this also fixes the Print button
being disabled only in the toolbar, but not in the secondary toolbar
(which is shown in smaller screens).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>