Prevent printing when download is hidden
Printing with the Print button from the toolbar as well as through
Ctrl|Meta+P uses the print service from PDFViewer, so when the download
is hidden the print service is now disabled.

When printing is flagged as not supported in the PDFViewerApplication
initialization the Print buttons are disabled, so it is not needed to
disable them explicity.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>