Allow downloads in sandboxed iframe
Until now it was possible to initiate a download from an iframe, but the
HTML spec has been updated to prevent that unless "allow-downloads" is
set in the "sandbox" attribute of the iframe.

As the PDF viewer shows the PDF file in an iframe the "allow-downloads"
attribute is needed to be able to download the PDF file by clicking on
the "Download" button of the UI in browsers that have implemented the
spec change (like Chromium 83; other browsers are expected to follow).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>