Disable content copy for PDF files that specify it
In v2.5.207 pdf.js added the "enablePermissions" option to disable
copying in the viewer for the PDF files that specify it. The option is
off by default, so it needs to be explicitly enabled.

Note that the content copy is prevented just by disabling selection
through CSS. The "selection disabled" flag in a PDF document is not a
strong security measure and can be easily circumvented (for example,
just downloading the file and opening it in a PDF reader that does not
honour the flag), so disabling the selection through CSS was deemed good
enough for pdf.js.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>