Fix footer not hidden when Talk sidebar is not shown
When the Talk sidebar is shown the footer in public pages is moved into
the "#app-content" element. However, by default the footer in public
pages is a direct child of the body, so this needs to be taken into
account when hiding the footer element, as the footer would not be moved
if Talk is not enabled.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>