Fix main content height when Talk sidebar is shown
The main content minimum height is set in server styles to the full
height minus the footer height. The PDF viewer hides the footer, so the
minimum height was increased to the full height to ensure that the main
content took all the available space.

However, when the Talk sidebar is shown, the page layout is also
adjusted, including the CSS styles. As the minimum height was set on the
element itself it took priority over the changes done by Talk, and
caused the PDF viewer to end past the window bottom when the Talk
sidebar is shown.

No matter if the Talk sidebar is shown or not the main content height
just needs to fill the available space, so now this is done using
flexbox. This avoids having to set an explicit height and thus works for
both the standard layout and the Talk sidebar layout. Moreover, even if
an explicit height is set somewhere, like in the server or Talk styles,
the flex-grow property implicitly overrides it and ensures that the main
content will fill the available height.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>