Do not load PDF viewer if public share has a download limit
The PDF viewer needs to download the file in order to render it in the
client. Due to this, if a public share has a download limit, loading the
PDF viewer automatically reduces the number of available downloads by
one, even if the user does not download the file manually. To prevent
that now the PDF viewer is not loaded in public shares if the share has
a download limit.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>