Replace deprecated "request" with "axios"
"request" was used only to get the "pdfjs-XXX-dist.zip" file when
building the app, but it was not used in the app itself.

It seems that it is not possible (with axios options only) to throttle
the "onDownloadProgress" event. Nevertheless, an external throttle
function was not used, as the event is not triggered very often and
therefore it does not need to be throttled.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>