Update "viewer.php" to PDF.js 3.9.179
The "viewer.php" acts as the "viewer.html" file included in the PDF.js
package. This commit updates "viewer.php" with the changes to
"viewer.html" from PDF.js 3.8.162 to 3.9.179 while respecting the
customized parts of "viewer.php".

This is needed to fix the initialization after bumping PDF.js to
3.9.179, as the new toolbar element, "editorStamp", seems to be expected
to be in the DOM and thus unconditionally used.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>