Fix extraction of PDF.js package
It seems that in some cases "unzipper" may not extract all the files
and, with some node versions, it may corrupt some of the extracted
files. As there is no fix for all that yet (downgrading "unzipper" to
0.10.11 might solve the missing extracted files, but it would still be
affected by the corrupted files and thus prevent upgrading to a newer
node version) "unzipper" was replaced by "adm-zip".

"adm-zip" does not have dependencies and has a large user base.
Unfortunately it does not support streams and requires loading the whole
file in memory to be able to extract it, but the package is not very
large (around 5,5 MiB), so this should not be a problem.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>