Use Viewer also in public pages for a single share
When the PDF viewer was opened in public pages for a single share the UI
was directly injected. Now the Viewer is open instead, which loads the
PDFView component and thus ensures that it will be used in all cases in
which the PDF viewer is used.

Note that a fallback was left to directly inject the UI if the Viewer
was not found, although this should not happen.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>