Import DAV helper functions from Viewer
The Viewer provides several useful helper functions for DAV, but they
can not be used (or at least I did not find out how) from the PDF
viewer. Therefore, for now they were just copied over.

The original file is written in TypeScript, but as it was not
extensively typed and TypeScript is not currently supported in the PDF
viewer for simplicity it was adjusted and converted to JavaScript.

The file was copied from "src/utils/davUtils.ts" of repository
"https://github.com/nextcloud/viewer" at commit
"854e63d33a8fd304a6c4ea6704f0f68c6308f4b8".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>