Enable "Save" button when there is an unsaved annotation
The "Save" button was always enabled, but saving and uploading the file
would be needed only if there is an unsaved annotation. Therefore now
the button is initially disabled, and it is enabled only when an
annotation is added; once the file is saved the button is disabled and
it will be enabled again once a new annotation is added.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>