Replace default "Download" icon with "Save" icon
The "Download / Save" button uses a "Download" icon by default. For
clarity, and to differentiate from the default PDF.js behaviour that
downloads the file with the added annotations rather than saving it in
place, the button now uses a "Save" icon instead.

Like other icons used in Nextcloud, the "Save" icon was copied from
Material Icons, which are licensed under the Apache License Version 2.0:
https://fonts.google.com/icons?selected=Material+Icons:save:

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>