Show forms as read only
Forms were shown as editable, but it was not possible to save them back
to Nextcloud (it will be possible in Nextcloud 28 and later). As the
form was editable anyway users could assume that the file was
automatically saved, leading to confusion when opening the file again
and finding that the changes were gone. To prevent that now the forms
are still shown, but only as read-only.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>