Show error when trying to open a shared PDF without download permissions
In order to show a PDF file it needs to be downloaded. Therefore, if a
shared PDF file does not have download permissions it is not possible to
show it (except for public shares, as in that case the download
permissions are not enforced), so now an error is shown instead.

The error is a custom one rather than a standard error from the viewer
(although with the same appearance) to better explain the reason.

Note that the error is shown only when the PDF file is loaded through
the viewer, which should be always the case. There is a fallback to
inject the UI in public shares in case the viewer is not available, but
as mentioned above download permissions are ignored in public shares and
that fallback should never be used anyway, so it was not taken into
account.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>