Register "share-attributes" DAV property in public share pages
The viewer loads the default and registered DAV properties in the
FileInfo object provided to the PDF viewer. The "share-attributes"
property is needed to know if downloads are allowed, but it is not one
of the properties included by default. The Files app registers it during
its initialization, so it is automatically available when the PDF viewer
is opened in the Files app. However, when the PDF viewer is opened in
public share page nothing registers it, so it needs to be explicitly
registered by the PDF viewer itself before the viewer is loaded.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>