Fix saving annotations in paths with special characters
The filename may contain characters that are not compatible with a URL
path, so it needs to be explicitly encoded. Otherwise the upload could
fail (for example, if the filename contains "%" followed by a letter,
which would cause the server to fail to decode the path and return a
"400 Bad request" error) or even overwrite a different file (for
example, if the filename contains "#", which would cause the rest of the
path to be ignored).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>