fix: Pin pdfjs-dist to exact 4.0.189
The PDF viewer explicitly sets "isEvalSupported" to "false", so it is
not affected by the security issue reported for versions <= 4.1.392,
which assume the default value of "true".

pdfjs-dist is the main dependency of the PDF viewer, and any version
update requires additional work in the PDF viewer, it is not just
increasing the version and that is it.

Due to all of the above, the pdfjs-dist version is pinned for now to
exact 4.0.189 to avoid dealing again and again with incorrect updates
after running "npm audit fix".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>