fix: Remove custom handling for public share pages with single files
Since Nextcloud 31 the public share pages for single files and folders
were unified, and public single file shares now show a file list with a
single file. Due to that it is no longer needed to explicitly inject the
PDF viewer in public shage pages or adjust the style, as the PDF viewer
will be loaded as needed by the Viewer component.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>