fix: Hide download based on the new WebDAV property
Since Nextcloud 31 the "hideDownload" HTML element no longer exists.
Between Nextcloud 31.0.0 and 31.0.3 (both included) "hide download"
behaved as "prevent download", so the PDF viewer could not show the file
at all in that case. In Nextcloud 31.0.4 the "hide download" behaviour
was restored, although the setting is now provided as a WebDAV property
(which is included in the file data), so the PDF viewer had to be
adjusted to that.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>