fix: Open PDF file in "richdocuments" if it is not downloadable
The PDF viewer must be able to download the file to show it, and if the
file is not downloadable an error message was just shown. However,
Nextcloud Office ("richdocuments") is able to show PDF files even if
they can not be downloaded, so now the viewer is reopened forcing
"richdocuments" to handle opening the file when it is not downloadable.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>
Signed-off-by: nextcloud-command <nextcloud-command@users.noreply.github.com>